fix: media upload
When install the module and use it, the images updated won't show up
because of the file permissions of the folder and files created, a way
to solved it to re-run the permissions fixes.

However instead of uploading directly into an new folder, will be simple
to use the media-library available into the InvoiceShelf.

With this in place, there is no need to re-run the permissions and
everything will be uploaded correctly.

Previously there is no way to remove the images uploaded, once defined
no change to removed, with that in mind, this issue was fixed by sending
a request informing the removal of the image.