chore: module auto packing
The artifact build was changed to include an extra WhiteLabel folder,
without that when unzipped the files will be placed directly into
Modules folder. The rsync was added to facilitated the files exclusion,
otherwise will require to set the full path of the files in zip command.

Artifact script still working locally since was introduced the build
folder.